Marker bug fix, file system tests

+++ app/fastapi_endpoint.py
@@ -1,6 +1,11 @@
+"""
+uvicorn fastapi_endpoint:app --host 0.0.0.0 --port 8000 --loop uvloop
+"""
+
 import asyncio
 import os
 from contextlib import asynccontextmanager
+from typing import Literal
 
 import aiofiles
 import magic
@@ -8,8 +13,10 @@ import uvloop
 from chat_utils.agent_backend import close_conn, get_agent_app
 from endpoint_utils import ALLOWED_IMAGE_TYPE, ALLOWED_PAPER_TYPE, respond_stream
 from fastapi import FastAPI, HTTPException, UploadFile
-from fastapi.responses import StreamingResponse
+from fastapi.responses import ORJSONResponse, StreamingResponse
 from fastapi.staticfiles import StaticFiles
+from file_utils.file_conversion import everything_to_markdown
+from pydantic import BaseModel
 
 asyncio.set_event_loop_policy(uvloop.EventLoopPolicy())
 
@@ -17,7 +24,6 @@ asyncio.set_event_loop_policy(uvloop.EventLoopPolicy())
 @asynccontextmanager
 async def lifespan(app: FastAPI):
     await get_agent_app()
-    print("agent app loaded")
     yield
     await close_conn()
 
@@ -28,7 +34,10 @@ app.mount("/media", StaticFiles(directory="media"), name="media")
 app.mount("/documents", StaticFiles(directory="documents"), name="documents")
 
 
-@app.post("/upload/image/{thread_id}")
+# 文件系统
+
+
+@app.post("/upload/image/{thread_id}", response_class=ORJSONResponse)
 async def upload_image(file: UploadFile, thread_id: str):
     # 类型验证
     chunk = await file.read(2048)
@@ -48,29 +57,40 @@ async def upload_image(file: UploadFile, thread_id: str):
     return {"url": final_path}
 
 
-@app.post("/upload/document/{user}/paper")
+@app.post("/upload/document/{user}/paper", response_class=ORJSONResponse)
 async def upload_paper(file: UploadFile, user: str):
+    # 重复上传保护
+    final_path = os.path.join("documents", user, "paper", file.filename)
+
     # 类型验证
     chunk = await file.read(2048)
 
     detected_mime = magic.from_buffer(chunk, mime=True)
-    if detected_mime not in ALLOWED_PAPER_TYPE:
-        raise HTTPException(
-            400,
-            detail=f"Only plain text, html, xml, epub, json, rtf, docx, xlsx, pptx, odt, pdf are allowed. Got {detected_mime}",
-        )
-    await file.seek(0)
+    if not os.path.exists(final_path):
+        if detected_mime not in ALLOWED_PAPER_TYPE:
+            raise HTTPException(
+                400,
+                detail=f"Only plain text, html, xml, epub, json, rtf, docx, xlsx, pptx, odt, pdf are allowed. Got {detected_mime}",
+            )
+        await file.seek(0)
 
-    # 异步流式上传
-    final_path = os.path.join("documents", user, "paper", file.filename)
-    async with aiofiles.open(final_path, "wb") as f:
-        while chunk := await file.read(8192):
-            await f.write(chunk)
+        # 异步流式上传
+        async with aiofiles.open(final_path, "wb") as f:
+            while chunk := await file.read(8192):
+                await f.write(chunk)
+    if detected_mime != "text/plain":
+        knowledgeBase_path = os.path.join("documents", user, "knowledgeBase")
+        await everything_to_markdown(final_path, knowledgeBase_path)
     return {"url": final_path}
 
 
-@app.post("/upload/document/{user}/code")
+@app.post("/upload/document/{user}/code", response_class=ORJSONResponse)
 async def upload_code(file: UploadFile, user: str):
+    # 重复上传保护
+    final_path = os.path.join("documents", user, "code", file.filename)
+    if os.path.exists(final_path):
+        return {"url": final_path}
+
     # 类型验证
     chunk = await file.read(2048)
 
@@ -83,13 +103,38 @@ async def upload_code(file: UploadFile, user: str):
     await file.seek(0)
 
     # 异步流式上传
-    final_path = os.path.join("documents", user, "code", file.filename)
     async with aiofiles.open(final_path, "wb") as f:
         while chunk := await file.read(8192):
             await f.write(chunk)
     return {"url": final_path}
 
 
-@app.post("/respond")
-async def respond(query: str, thread_id: str, chat_mode: str, current_user: str):
-    return StreamingResponse(respond_stream(query, thread_id, chat_mode, current_user))
+@app.get("/list_files/{user}/{directory}", response_class=ORJSONResponse)
+async def list_directory(
+    user: str,
+    directory: Literal["paper", "knowledgeBase", "code", "tempest", "convert"],
+):
+    dir_path = os.path.join("documents", user, directory)
+    with os.scandir(dir_path) as entries:
+        return [entry.name for entry in entries]
+
+
+# 模型回复
+
+
+class ChatQuery(BaseModel):
+    query: str
+    chat_mode: Literal["常规", "工具", "多模态", "知识库", "网页搜索"]
+    current_user: str
+
+
+@app.post("/chat/{thread_id}")
+async def respond(thread_id: str, chat_query: ChatQuery):
+    return StreamingResponse(
+        respond_stream(
+            chat_query.query,
+            thread_id,
+            chat_query.chat_mode,
+            chat_query.current_user,
+        )
+    )
